# public/api-docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "youmatter API Documentation",
    "version": "1.0.0",
    "description": "Comprehensive REST API documentation for youmatter English Learning Platform",
    "contact": {
      "name": "youmatter Support",
      "email": "support@youmatter.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://youmatter.com",
      "description": "Production server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/api/admin/analytics": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Get platform analytics",
        "description": "System admin endpoint for platform metrics and statistics",
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Platform analytics data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "total_users": {
                      "type": "integer"
                    },
                    "total_patients": {
                      "type": "integer"
                    },
                    "total_therapists": {
                      "type": "integer"
                    },
                    "total_lessons": {
                      "type": "integer"
                    },
                    "total_resources": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized - System admin only"
          }
        }
      }
    }
  }
}